requirements.txt
python-telegram-bot==13.15
python-dotenv==1.0.0
mercadopago==2.2.0
apscheduler==3.6.3
urllib3==1.26.15
six==1.15.0 